+++ src/java.base/share/classes/jdk/internal/org/objectweb/asm/MethodWriter.java
@@ -1245,7 +1245,7 @@ final class MethodWriter extends MethodVisitor {
                     // one place, but this does not work for labels which have not been visited yet.
                     // Therefore, when we detect here two labels having the same bytecode offset, we need to
                     // - consolidate the state scattered in these two instances into the canonical instance:
-                    currentBasicBlock.flags |= (label.flags & Label.FLAG_JUMP_TARGET);
+                    currentBasicBlock.flags |= (short)(label.flags & Label.FLAG_JUMP_TARGET);
                     // - make sure the two instances share the same Frame instance (the implementation of
                     // {@link Label#getCanonicalInstance} relies on this property; here label.frame should be
                     // null):
@@ -1261,7 +1261,7 @@ final class MethodWriter extends MethodVisitor {
             if (lastBasicBlock != null) {
                 if (label.bytecodeOffset == lastBasicBlock.bytecodeOffset) {
                     // Same comment as above.
-                    lastBasicBlock.flags |= (label.flags & Label.FLAG_JUMP_TARGET);
+                    lastBasicBlock.flags |= (short)(label.flags & Label.FLAG_JUMP_TARGET);
                     // Here label.frame should be null.
                     label.frame = lastBasicBlock.frame;
                     currentBasicBlock = lastBasicBlock;
